refactor: extract rendering from transistor boss  class for better readability and SoC
